% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{This function is not intended for use by the user. It is called by the commands and serves as an interpreter from R to FNA.}
\usage{
check(match_call_list)
}
\arguments{
\item{match_call_list}{: FNA command match list}
}
\value{
: Commmand in the form accepted by FNA.
}
\description{
This function is not intended for use by the user. It is called by the commands and serves as an interpreter from R to FNA.
}
