% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Downloads a csv file from a FNA workspace}
\usage{
download(login_token, workspace, file)
}
\arguments{
\item{workspace}{Workspace where the file is locatd}

\item{file}{Name of the file to download}

\item{login_data}{Generated by the login function}
}
\value{
dataframe containing the downloaded file
}
\description{
Downloads a csv file from a FNA workspace
}
