% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_command.R
\name{exec_command}
\alias{exec_command}
\title{Execute a command in FNA in the workspace defined by the set_workspace command. This function is used by custom functions in this package.}
\usage{
exec_command(script)
}
\arguments{
\item{script}{: FNA script to be executed}
}
\value{
: None
}
\description{
Execute a command in FNA in the workspace defined by the set_workspace command. This function is used by custom functions in this package.
}
\examples{
: exec_command('arc.property.define -property color -type Color')
}
