% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\title{Lists files in a given workspace}
\usage{
files(workspace)
}
\arguments{
\item{workspace}{: the workspace in which the files will be located.}

\item{login_token}{: generated with the login function.}
}
\value{
: dataframe containing of the file names in the given directory
}
\description{
Lists files in a given workspace
}
