% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commands.R
\name{get_commands}
\alias{get_commands}
\title{Load all commands available on the FNA platform}
\usage{
get_commands()
}
\arguments{
\item{login_data}{}
}
\value{
The list of commands available to the user with the specified user token.
}
\description{
Load all commands available on the FNA platform
}
