% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Logs into fna}
\usage{
login(token, username, base_url)
}
\arguments{
\item{token}{: FNA Application token (You can find the token here https://www.fnalab.com/settings#application)}

\item{username}{: FNA username}

\item{base_url}{: URL of the FNA instance}
}
\value{
list of login informartion
}
\description{
Logs into fna
}
