% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_workspace.R
\name{set_workspace}
\alias{set_workspace}
\title{Moves to the workspace chosen by the user. Note that the function does not perform any validation - users are advised to use first the function workspaces to load all workspaces.}
\usage{
set_workspace(workspace)
}
\arguments{
\item{workspace}{: A valid workspace on FNA.}
}
\value{
: None
}
\description{
Moves to the workspace chosen by the user. Note that the function does not perform any validation - users are advised to use first the function workspaces to load all workspaces.
}
