% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload}
\alias{upload}
\title{Uploads a dataframe R to FNA}
\usage{
upload(login_data, workspace, df)
}
\arguments{
\item{login_data}{generated with the login function.}

\item{workspace}{workspace to which the dataframe should be uploaded}

\item{df}{dataframe}
}
\value{
Returns whether the upload was successful
}
\description{
Uploads a dataframe R to FNA
}
