% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{workspaces}
\alias{workspaces}
\title{:Lists workspaces available for a user}
\usage{
workspaces()
}
\arguments{
\item{login_token}{}
}
\value{
dataframe listing the workspaces for a user
}
\description{
:Lists workspaces available for a user
}
